backend/requirements.txt

Flask==3.0.3
Flask-Cors==6.0.2
Flask-Limiter==3.5.1
requests==2.32.5
groq==0.11.0
reportlab==4.4.10
python-dotenv==1.2.2
gunicorn==25.1.0
